removed executable flag from regular files; fixed README to look nicer on github; move version to separate file to safely import it from setup.py
