+++ README
@@ -1,1 +1,1 @@
-README.org
+README.rst
