+++ py.test.integrate_with_distutils.help.rst
